refactor: move navGroup related interface into a service
