+++ src/core/public/chrome/index.ts
@@ -36,7 +36,6 @@ export {
   ChromeStart,
   InternalChromeStart,
   ChromeHelpExtension,
-  ChromeRegistrationNavLink,
 } from './chrome_service';
 export {
   ChromeHelpExtensionMenuLink,
@@ -51,3 +50,4 @@ export { ChromeRecentlyAccessed, ChromeRecentlyAccessedHistoryItem } from './rec
 export { ChromeNavControl, ChromeNavControls } from './nav_controls';
 export { ChromeDocTitle } from './doc_title';
 export { RightNavigationOrder } from './constants';
+export { ChromeRegistrationNavLink } from './nav_group';
