+++ src/core/public/chrome/nav_group/index.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export {
+  ChromeNavGroupService,
+  ChromeNavGroupServiceSetupContract,
+  ChromeNavGroupServiceStartContract,
+  ChromeRegistrationNavLink,
+  NavGroupItemInMap,
+} from './nav_group_service';
